+++ packages/docs/angular-table-docs/src/routes/features+/basic+/demos/basic-demo.ts
@@ -19,8 +19,10 @@ import {type User, userColumns, users} from "./data"
             ) {
               <tr q-table-row>
                 @for (header of headerGroup.headers; track header.id) {
-                  <th *renderHeader="header; let value" q-table-header-cell>
-                    {{ value }}
+                  <th q-table-header-cell>
+                    <ng-container *renderHeader="header; let value">
+                      {{ value }}
+                    </ng-container>
                   </th>
                 }
               </tr>
@@ -30,8 +32,10 @@ import {type User, userColumns, users} from "./data"
             @for (row of table.getRowModel().rows; track row.id) {
               <tr q-table-row>
                 @for (cell of row.getVisibleCells(); track cell.id) {
-                  <td *renderCell="cell; let value" q-table-cell>
-                    {{ value }}
+                  <td q-table-cell>
+                    <ng-container *renderCell="cell; let value">
+                      {{ value }}
+                    </ng-container>
                   </td>
                 }
               </tr>