+++ packages/docs/react-table-docs/src/routes/features+/filters-client-side+/demos/index.ts
@@ -1,0 +1,3 @@
+export * from "./filters-client-side-demo"
+export * from "./filters"
+export * from "./use-data"